+++ src/SoundGame.cpp
@@ -2,9 +2,9 @@
 #include "../lib/SoundGame.h"
 
 uint16_t SoundGame::play() {
-    uint16_t previousMillis = millis();
+    unsigned long previousMillis = millis();
 
-    uint16_t interval = beginGame() + previousMillis;
+    unsigned long interval = beginGame() + previousMillis;
 
     while (true) {
         if (millis() > interval) {
